IndexOutOfRange uses ints to handle -1, etc.

+++ OpenSim/Common/Exception.h
@@ -172,9 +172,9 @@ public:
     IndexOutOfRange(const std::string& file,
                     size_t line,
                     const std::string& func,
-                    size_t index, 
-                    size_t min, 
-                    size_t max) :
+                    int index,
+                    int min, 
+                    int max) :
         Exception(file, line, func) {
         std::string msg = "min = " + std::to_string(min);
         msg += " max = " + std::to_string(max);
